chore: move the logic for converting invalid utf8 strings to jtrace package
