+++ internal/jptrace/utf8.go
@@ -1,4 +1,4 @@
-package utils
+package jptrace
 
 import (
 	"golang.org/x/text/runes"
